requirements.txt
# Core numerical and data handling
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine learning & statistics
scikit-learn>=1.3.0
xgboost>=1.7.6
lightgbm>=3.3.5
umap-learn>=0.5.4
joblib>=1.3.0

# Deep learning (select the framework you need)
tensorflow>=2.13.0
torch>=2.0.0

# Model serving & dashboards
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
streamlit>=1.25.0
dash>=2.11.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=1.10.0
pyyaml>=6.0.0

# Tooling & tests
pytest>=7.4.0
httpx>=0.26.0
